+++ AuthPermissions/BulkLoadServices/IBulkLoadTenantsService.cs
@@ -1,8 +1,9 @@
 ﻿// Copyright (c) 2021 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
+using AuthPermissions.SetupCode;
 using StatusGeneric;
 
 namespace AuthPermissions.BulkLoadServices
@@ -13,20 +14,13 @@ namespace AuthPermissions.BulkLoadServices
     public interface IBulkLoadTenantsService
     {
         /// <summary>
-        /// This allows you to define tenants in a bulk load from a string. Each line in that string should hold a tenant
-        /// (a line is ended with <see cref="Environment.NewLine"/>)
-        /// If you are using a hierarchical tenant design, then you must define the higher company first
+        /// This allows you to add tenants to the database on startup.
+        /// It gets the definition of each tenant from the <see cref="BulkLoadTenantDto"/> class
         /// </summary>
-        /// <param name="linesOfText">If you are using a single layer then each line contains the a tenant name
-        /// If you are using hierarchical tenant, then each line contains the whole hierarchy with '|' as separator, e.g.
-        /// Holding company
-        /// Holding company | USA branch 
-        /// Holding company | USA branch | East Coast 
-        /// Holding company | USA branch | East Coast | Washington
-        /// Holding company | USA branch | East Coast | NewYork
+        /// <param name="tenantSetupData">If you are using a single layer then each line contains the a tenant name
         /// </param>
         /// <param name="options">The AuthPermissionsOptions to check what type of tenant setting you have</param>
         /// <returns></returns>
-        Task<IStatusGeneric> AddTenantsToDatabaseAsync(string linesOfText, AuthPermissionsOptions options);
+        Task<IStatusGeneric> AddTenantsToDatabaseAsync(List<BulkLoadTenantDto> tenantSetupData, AuthPermissionsOptions options);
     }
 }