+++ AuthPermissions/CommonCode/ErrorReportingExtensions.cs
@@ -76,7 +76,7 @@ namespace AuthPermissions.CommonCode
         public static string FormErrorString(this string line, int lineNum, int charNum, string error)
         {
             var charPart = charNum < 0 ? "" : $", char: {charNum + 1}";
-            var result = $"Line/index {lineNum + 1:####}{charPart}: {error}{Environment.NewLine}{line}";
+            var result = $"Index {lineNum + 1:####}{charPart}: {error}{Environment.NewLine}{line}";
             if (charNum > -1)
                 result += Environment.NewLine + new string(' ', charNum) + "|";