+++ Test/UnitTests/TestExamples/TestExample4AdminDisplayUpdate.cs
@@ -37,7 +37,7 @@ namespace Test.UnitTests.TestExamples
                 })
                 .UsingInMemoryDatabase()
                 .AddRolesPermissionsIfEmpty(Example4AppAuthSetupData.RolesDefinition)
-                .AddTenantsIfEmpty(Example4AppAuthSetupData.BulkHierarchicalTenants)
+                .AddTenantsIfEmpty(Example4AppAuthSetupData.TenantDefinition.ToList()) //HAVE TO TAKE A COPY!!!
                 .AddAuthUsersIfEmpty(Example4AppAuthSetupData.UsersRolesDefinition)
                 .RegisterFindUserInfoService<StubIFindUserInfoFactory.StubIFindUserInfo>()
                 .SetupForUnitTestingAsync();